Backport ReaderOptions exposition in ez-rpc
After consulting with the capnproto devs, I've sent them a pull request
which will allow ReaderOptions to be optionally passed into ez-rpc. As
trusty will obviously not be shipping this any time soon, even if
accepted tomorrow, I've included the modified files into this
repository. These files are under MIT as well, but not under our
copyright.

In the long term, when trusty gets these changes, we should intend to
remove all files in the modproto directory.